requirements.txt
asgiref==3.2.7
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
cycler==0.10.0
decorator==4.4.2
Django==3.0.7
Flask==1.1.2
Flask-WTF==0.14.3
idna==2.10
imageio==2.9.0
itsdangerous==1.1.0
Jinja2==2.11.2
kiwisolver==1.3.1
MarkupSafe==1.1.1
matplotlib==3.3.3
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.1
PyWavelets==1.1.1
requests==2.25.1
scikit-image==0.18.1
scipy==1.5.4
six==1.15.0
sqlparse==0.3.1
tifffile==2020.12.8
urllib3==1.26.2
Werkzeug==1.0.1
WTForms==2.3.1
